Go: add whitespace attachment expectation to all rewriteRun tests
