+++ rewrite-go/pkg/printer/whitespace_attachment_test.go
@@ -17,113 +17,17 @@
 package printer_test
 
 import (
-	"fmt"
-	"reflect"
 	"strings"
 	"testing"
 
 	"github.com/openrewrite/rewrite/rewrite-go/pkg/parser"
-	"github.com/openrewrite/rewrite/rewrite-go/pkg/printer"
-	"github.com/openrewrite/rewrite/rewrite-go/pkg/tree/java"
-	"github.com/openrewrite/rewrite/rewrite-go/pkg/visitor"
+	"github.com/openrewrite/rewrite/rewrite-go/pkg/test"
 )
 
-// This mirrors rewrite-javascript's whitespace-attachment.test.ts: leading
-// whitespace must be attached to the outermost LST element rather than leaking
-// onto the prefix of a node's first child. If a node emits nothing of its own
-// before its first child, and that child carries a non-empty whitespace prefix,
-// the whitespace should have been attached to the parent instead.
-//
-// The JS test subclasses the printer to override the virtual beforeSyntax /
-// afterSyntax hooks and build a node tree. The Go printer routes prefixes
-// through per-method beforeSyntax calls that are not virtually dispatched, but
-// PreVisit *is* dispatched via the Self field and fires for every node right
-// before it emits its prefix. Recording the output-buffer length at each
-// PreVisit (the node's start offset) lets us reconstruct, for every parent,
-// whether it emitted anything before its first child — which is exactly what
-// the JS test checks by looking at children[0].
-
-// captureNode is one node in the reconstructed print tree.
-type captureNode struct {
-	kind             string
-	startOffset      int    // output length when this node started printing
-	prefixWhitespace string // the node's own prefix whitespace, emitted first
-	hasPrefix        bool   // whether the node is a J carrying a prefix
-	children         []*captureNode
-}
-
-// treeCapturingPrinter embeds the real GoPrinter and overrides PreVisit to
-// record the print tree without altering any printing behavior.
-type treeCapturingPrinter struct {
-	printer.GoPrinter
-	byTree map[java.Tree]*captureNode
-	roots  []*captureNode
-}
-
-func (p *treeCapturingPrinter) PreVisit(t java.Tree, param any) java.Tree {
-	out := param.(*printer.PrintOutputCapture)
-
-	node := &captureNode{
-		kind:        prettifyKind(t),
-		startOffset: len(out.String()),
-	}
-	if j, ok := t.(java.J); ok {
-		node.hasPrefix = true
-		node.prefixWhitespace = j.GetPrefix().Whitespace
-	}
-	p.byTree[t] = node
-
-	if parent := p.Cursor().Parent(); parent != nil && parent.Value() != nil {
-		if pn, ok := p.byTree[parent.Value()]; ok {
-			pn.children = append(pn.children, node)
-		} else {
-			p.roots = append(p.roots, node)
-		}
-	} else {
-		p.roots = append(p.roots, node)
-	}
-
-	return t
-}
-
-// prettifyKind renders a node's type as "package.TypeName" (e.g.
-// "java.MethodInvocation", "golang.Composite").
-func prettifyKind(t java.Tree) string {
-	name := reflect.TypeOf(t).String()
-	return strings.TrimPrefix(name, "*")
-}
-
-func findWhitespaceViolations(roots []*captureNode) []string {
-	var violations []string
-
-	var check func(n *captureNode)
-	check = func(n *captureNode) {
-		if len(n.children) > 0 {
-			first := n.children[0]
-			// first.startOffset == n.startOffset means the parent emitted
-			// nothing before its first child — i.e. the child is the parent's
-			// children[0] (equivalent to firstChild being a node in the JS test).
-			if first.startOffset == n.startOffset && first.hasPrefix {
-				ws := first.prefixWhitespace
-				if ws != "" && strings.TrimSpace(ws) == "" {
-					violations = append(violations, fmt.Sprintf(
-						"%s has child %s starting with whitespace |%s|. "+
-							"The whitespace should rather be attached to %s.",
-						n.kind, first.kind, ws, n.kind))
-				}
-			}
-		}
-		for _, c := range n.children {
-			check(c)
-		}
-	}
-
-	for _, r := range roots {
-		check(r)
-	}
-	return violations
-}
-
+// Leading whitespace must be attached to the outermost LST element rather than
+// leaking onto the prefix of a node's first child. The enforcement logic lives
+// in test.WhitespaceAttachmentViolations and is also asserted against the
+// whole native Go test corpus from the RewriteRun harness.
 func TestWhitespaceAttachedToOutermostElement(t *testing.T) {
 	sources := []string{
 		"package main\n\nfunc f() int { return 1 + 2 }\n",
@@ -147,16 +51,11 @@ func TestWhitespaceAttachedToOutermostElement(t *testing.T) {
 			if err != nil {
 				t.Fatalf("parse: %v", err)
 			}
-			capturing := visitor.Init(&treeCapturingPrinter{
-				byTree: make(map[java.Tree]*captureNode),
-			})
-			out := printer.NewPrintOutputCaptureWithMarkers(printer.SanitizedMarkerPrinter)
 
 			// when
-			capturing.Visit(cu, out)
+			violations := test.WhitespaceAttachmentViolations(cu)
 
 			// then
-			violations := findWhitespaceViolations(capturing.roots)
 			if len(violations) > 0 {
 				t.Errorf("expected no whitespace-attachment violations, got %d:\n%s",
 					len(violations), strings.Join(violations, "\n"))
